+++ Changelog.md
@@ -75,6 +75,9 @@
     * [Issue #1789](https://bitbucket.org/osrf/gazebo/issue/1789)
     * [Pull request #2019](https://bitbucket.org/osrf/gazebo/pull-request/2019)
 
+1 Implemented transport clear buffers
+  * [Pull request #2017](https://bitbucket.org/osrf/gazebo/pull-request/2017)
+
 1. KeyEvent constructor should be in a source file. Removed a few visibility
 flags from c functions. Windows did not like `CPPTYPE_*` in
 `gazebo/gui/ConfigWidget.cc`, so I replaced it with `TYPE_*`.